Created function to find smallest tree
Function is much slower, but is (almost) guaranteed to return the tree with the smallest amount of intermediate elements and crafts.

The one case where it isn't (and this is entirely theoretical, have not been able to find an example), is where it actually  requires less intermediate elements/crafts to backtrack to a certain required element rather than making it from scratch. Not even sure if this can happen, but my current algorithm only searches the predepth recipes (ones that use lower depth items than the target).